Inital commit + boost grab function
This is the inital commit from the sample RL bot code, as well as some initial work on the `get_all_boosts` function in `/src/bot.py` file.

Currently the bot starts on blue team, and drives over the boosts in order to pick up the boosts, this whole process takes 28 seconds to get back to the same boost.

I'd like to speed this up with better turning and boost usage, maybe adding in some speed flips (because why not), and then it'd be fun to make it so when you trigger this function it'll start with the closest boost and work around from there.