fix get_kickoff_position
Previously this method always returned 5 because comparing two objects with `==` checks to see if they are the same object, not if the contents are the same. This also probably wouldn't have worked because car_location has a z parameter and the Vec3's I was passing in only had an X and Y.

Should I want to go down the road of comparing objects in the future, I should put a `` def __eq__(self, value):`` on the object that will consider equality by the the contents of the object.

The solution implemented here works best since we're flattening the location (setting z to 0) and then measuring the distance from the car_location object to provided Vec3. This should be 0 if we're in the right place, so checking if its < 1 gives us some quantum wiggle room.